requirements.txt
google-adk==1.2.1
python-dotenv
toolbox-core 
langchain
google-genai
a2a-sdk
google-cloud-aiplatform[adk,agent_engines]==1.96
locust
google-cloud-spanner==3.54.0
google-adk[eval]
python-a2a